pyproject.toml
[project]
name = "ai-resume"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "markitdown>=0.0.1a3",
    "pdflatex>=0.1.3",
    "pydantic-ai>=0.0.16",
    "python-dotenv>=1.0.1",
    "pyairtable>=2.3.3",
]
